+++ workspaces/quay/plugins/quay/CHANGELOG.md
@@ -1,5 +1,11 @@
 ### Dependencies
 
+## 1.14.4
+
+### Patch Changes
+
+- a6e850f: Updated dependency `msw` to `1.3.5`.
+
 ## 1.14.3
 
 ### Patch Changes