+++ workspaces/tekton/plugins/tekton/CHANGELOG.md
@@ -1,5 +1,12 @@
 ### Dependencies
 
+## 3.16.2
+
+### Patch Changes
+
+- a6e850f: Updated dependency `msw` to `1.3.5`.
+- d3d078b: Updated dependency `@types/lodash` to `4.17.13`.
+
 ## 3.16.1
 
 ### Patch Changes